requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
line-bot-sdk==3.13.0
pydantic==2.8.2
openai==1.37.0
